Move recordc sort code into its own file
